// src/modules/utils/books/enum/genre-query.enum.ts
enum EGenreQuery {
  ALL = 'all',
  LITERARY_FICTION = 'Literary_Fiction',
  MYSTERY = 'Mystery',
  THRILLER = 'Thriller',
  HORROR = 'Horror',
  HISTORICAL = 'Historical',
  ROMANCE = 'Romance',
  WESTERN = 'Western',
  BILDUNGSROMAN = 'Bildungsroman',
  SPECULATIVE_FICTION = 'Speculative_Fiction',
}

export default EGenreQuery